Control.Monad.IOSim

Simulation monad

data IOSim s a

type STMSim = STM

Run simulation

runSim ∷ ∀ a. (∀ s. IOSim s a) → Either Failure a

IOSim is a pure monad.

runSimOrThrow ∷ ∀ a. (∀ s. IOSim s a) → a

For quick experiments and tests it is often appropriate and convenient to simply throw failures as exceptions.

runSimStrictShutdown ∷ ∀ a. (∀ s. IOSim s a) → Either Failure a

Like runSim but also fail if when the main thread terminates, there are other threads still running or blocked. If one is trying to follow a strict thread cleanup policy then this helps testing for that.

data Failure

Simulation termination with failure

Constructors

FailureException SomeException
The main thread terminated with an exception

FailureDeadlock ![Labelled ThreadId]
The threads all deadlocked

FailureSloppyShutdown [Labelled ThreadId]
The main thread terminated normally but other threads were still alive, and strict shutdown checking was requested. See runSimStrictShutdown

runSimTrace ∷ ∀ a. (∀ s. IOSim s a) → SimTrace a

See runSimTraceST below.

controlSimTrace

Arguments

∷ ∀ a. Maybe Int

→ ScheduleControl
note: must be either ControlDefault or ControlAwait.

→ (∀ s. IOSim s a)

→ SimTrace a

exploreSimTrace ∷ ∀ a test. Testable test ⇒ (ExplorationOptions → ExplorationOptions) → (∀ s. IOSim s a) → (Maybe (SimTrace a) → SimTrace a → test) → Property

data ScheduleControl

Constructors

ControlDefault
default scheduling mode

ControlAwait [ScheduleMod]
if the current control is ControlAwait, the normal scheduling will proceed, until the thread found in the first ScheduleMod reaches the given step. At this point the thread is put to sleep, until after all the steps are followed.

ControlFollow [StepId] [ScheduleMod]
follow the steps then continue with schedule modifications. This control is set by followControl when controlTargets returns true.

runSimTraceST ∷ ∀ s a. IOSim s a → ST s (SimTrace a)

The most general method of running IOSim is in ST monad. One can recover failures or the result from SimTrace with traceResult, or access SimEventTypes generated by the computation with traceEvents. A slightly more convenient way is exposed by runSimTrace.

liftST ∷ ST s a → IOSim s a

traceM ∷ Typeable a ⇒ a → IOSim s ()

traceSTM ∷ Typeable a ⇒ a → STMSim s ()

Simulation time

setCurrentTime ∷ UTCTime → IOSim s ()

Set the current wall clock time for the thread's clock domain.

unshareClock ∷ IOSim s ()

Put the thread into a new wall clock domain, not shared with the parent thread. Changing the wall clock time in the new clock domain will not affect the other clock of other threads. All threads forked by this thread from this point onwards will share the new clock domain.

Simulation trace

type SimTrace a = Trace (SimResult a) SimEvent

data Trace a b

A cons list with polymorphic nil, thus an octopus.

Trace Void a is an infinite stream

Trace () a is isomorphic to [a]

Usually used with a being a non empty sum type.

Constructors

Cons b (Trace a b)

Nil a

Bundled Patterns

pattern Trace ∷ Time → ThreadId → Maybe ThreadLabel → SimEventType → SimTrace a → SimTrace a

Deprecated: Use SimTrace instead.

pattern SimTrace ∷ Time → ThreadId → Maybe ThreadLabel → SimEventType → SimTrace a → SimTrace a

pattern TraceDeadlock ∷ Time → [Labelled ThreadId] → SimTrace a

pattern TraceLoop ∷ SimTrace a

pattern TraceMainReturn ∷ Time → a → [Labelled ThreadId] → SimTrace a

pattern TraceMainException ∷ Time → SomeException → [Labelled ThreadId] → SimTrace a

pattern TraceRacesFound ∷ [ScheduleControl] → SimTrace a → SimTrace a

data SimResult a

Constructors

MainReturn !Time a ![Labelled ThreadId]

MainException !Time SomeException ![Labelled ThreadId]

Deadlock !Time ![Labelled ThreadId]

Loop

data SimEvent

Trace is a recursive data type, it is the trace of a IOSim computation. The trace will contain information about thread sheduling, blocking on TVars, and other internal state changes of IOSim. More importantly it also supports traces generated by the computation with say (which corresponds to using putStrLn in IO), traceEventM, or dynamically typed traces with traceM (which generalise the base library traceM)

It also contains information on races discovered.

See also: traceEvents, traceResult, selectTraceEvents, selectTraceEventsDynamic and printTraceEventsSay.

Constructors

SimEvent

Fields

seTime ∷ !Time

seThreadId ∷ !ThreadId

seThreadLabel ∷ !(Maybe ThreadLabel)

seType ∷ !SimEventType

SimRacesFound [ScheduleControl]

data SimEventType

Constructors

EventSay String

EventLog Dynamic

EventMask MaskingState

EventThrow SomeException

EventThrowTo SomeException ThreadId

EventThrowToBlocked

EventThrowToWakeup

EventThrowToUnmasked (Labelled ThreadId)

EventThreadForked ThreadId

EventThreadFinished

EventThreadUnhandled SomeException

EventTxCommitted [Labelled TVarId] [TVarId]

EventTxAborted

EventTxBlocked [Labelled TVarId]

EventTxWakeup [Labelled TVarId]

EventTimerCreated TimeoutId TVarId Time

EventTimerUpdated TimeoutId Time

EventTimerCancelled TimeoutId

EventTimerExpired TimeoutId

EventThreadSleep

EventThreadWake

type ThreadLabel = String

data Labelled a

Constructors

Labelled

Fields

l_labelled ∷ !a

l_label ∷ !(Maybe String)

Pretty printers

ppTrace ∷ Show a ⇒ SimTrace a → String

Pretty print simulation trace.

ppTrace_ ∷ SimTrace a → String

Like ppTrace but does not show the result value.

ppEvents ∷ [(Time, ThreadId, Maybe ThreadLabel, SimEventType)] → String

ppSimEvent

Arguments

∷ Int
width of thread label

→ SimEvent

→ String

Selectors

traceEvents ∷ SimTrace a → [(Time, ThreadId, Maybe ThreadLabel, SimEventType)]

traceResult ∷ Bool → SimTrace a → Either Failure a

list selectors

selectTraceEvents ∷ (SimEventType → Maybe b) → SimTrace a → [b]

selectTraceEvents' ∷ (SimEventType → Maybe b) → SimTrace a → [b]

selectTraceEventsDynamic ∷ ∀ a b. Typeable b ⇒ SimTrace a → [b]

Select all the traced values matching the expected type. This relies on the sim's dynamic trace facility.

For convenience, this throws exceptions for abnormal sim termination.

selectTraceEventsDynamic' ∷ ∀ a b. Typeable b ⇒ SimTrace a → [b]

Like selectTraceEventsDynamic but returns partial trace if an exception is found in it.

selectTraceEventsSay ∷ SimTrace a → [String]

Get a trace of EventSay.

For convenience, this throws exceptions for abnormal sim termination.

selectTraceEventsSay' ∷ SimTrace a → [String]

Like selectTraceEventsSay but return partial trace if an exception is found in it.

selectTraceRaces ∷ SimTrace a → [ScheduleControl]

trace selectors

traceSelectTraceEvents ∷ (SimEventType → Maybe b) → SimTrace a → Trace (SimResult a) b

The most general select function. It is a _total_ function.

traceSelectTraceEventsDynamic ∷ ∀ a b. Typeable b ⇒ SimTrace a → Trace (SimResult a) b

Select dynamic events. It is a _total_ function.

traceSelectTraceEventsSay ∷ ∀ a. SimTrace a → Trace (SimResult a) String

Select say events. It is a _total_ function.

IO printer

printTraceEventsSay ∷ SimTrace a → IO ()

Print all EventSay to the console.

For convenience, this throws exceptions for abnormal sim termination.

Exploration options

type ExplorationSpec = ExplorationOptions → ExplorationOptions

data ExplorationOptions

stdExplorationOptions ∷ ExplorationOptions

withScheduleBound ∷ Int → ExplorationSpec

withBranching ∷ Int → ExplorationSpec

withStepTimelimit ∷ Int → ExplorationSpec

withReplay ∷ ScheduleControl → ExplorationSpec

Eventlog

newtype EventlogEvent

Wrapper for Eventlog events so they can be retrieved from the trace with selectTraceEventsDynamic.

Constructors

EventlogEvent String

newtype EventlogMarker

Wrapper for Eventlog markers so they can be retrieved from the trace with selectTraceEventsDynamic.

Constructors

EventlogMarker String

Low-level API

execReadTVar ∷ TVar s a → ST s a

Deprecated interfaces

type SimM s = IOSim s

Deprecated: Use IOSim

type SimSTM = STM

Deprecated: Use STMSim

type TraceEvent = SimEventType

Deprecated: Use SimEventType instead.