Control.Monad.IOSim.CommonTypes

Description

Common types shared between IOSim and IOSimPOR.

Documentation

data ThreadId

Constructors

RacyThreadId [Int]

ThreadId [Int]

childThreadId ∷ ThreadId → Int → ThreadId

setRacyThread ∷ ThreadId → ThreadId

newtype TVarId

Constructors

TVarId Int

newtype TimeoutId

Constructors

TimeoutId Int

newtype ClockId

Constructors

ClockId [Int]

newtype VectorClock

Constructors

VectorClock

Fields

getVectorClock ∷ Map ThreadId Int

unTimeoutId ∷ TimeoutId → Int

type ThreadLabel = String

type TVarLabel = String

data TVar s a

Constructors

TVar

Fields

tvarId ∷ !TVarId

The identifier of this var.

tvarLabel ∷ !(STRef s (Maybe TVarLabel))

Label.

tvarCurrent ∷ !(STRef s a)

The var's current value

tvarUndo ∷ !(STRef s [a])

A stack of undo values. This is only used while executing a transaction.

tvarBlocked ∷ !(STRef s ([ThreadId], Set ThreadId))

Thread Ids of threads blocked on a read of this var. It is represented in reverse order of thread wakeup, without duplicates.

To avoid duplicates efficiently, the operations rely on a copy of the thread Ids represented as a set.

tvarVClock ∷ !(STRef s VectorClock)

The vector clock of the current value.

tvarTrace ∷ !(STRef s (Maybe (Maybe a → a → ST s TraceValue)))

Callback to construct a trace which will be attached to the dynamic trace.

data SomeTVar s where

Constructors

SomeTVar ∷ !(TVar s a) → SomeTVar s

data Deschedule

Constructors

Yield

Interruptable

Blocked

Terminated

Sleep