Control.Monad.IOSim.Internal

Documentation

newtype IOSim s a

Constructors

IOSim

Fields

unIOSim ∷ ∀ r. (a → SimA s r) → SimA s r

type SimM s = IOSim s

Deprecated: Use IOSim

runIOSim ∷ IOSim s a → SimA s a

runSimTraceST ∷ ∀ s a. IOSim s a → ST s (Trace a)

The most general method of running IOSim is in ST monad. One can recover failures or the result from Trace with traceResult, or access TraceEvents generated by the computation with traceEvents. A slightly more convenient way is exposed by runSimTrace.

traceM ∷ Typeable a ⇒ a → IOSim s ()

traceSTM ∷ Typeable a ⇒ a → STMSim s ()

data STM s a

type STMSim = STM

type SimSTM = STM

Deprecated: Use STMSim

setCurrentTime ∷ UTCTime → IOSim s ()

Set the current wall clock time for the thread's clock domain.

unshareClock ∷ IOSim s ()

Put the thread into a new wall clock domain, not shared with the parent thread. Changing the wall clock time in the new clock domain will not affect the other clock of other threads. All threads forked by this thread from this point onwards will share the new clock domain.

newtype TimeoutException

Constructors

TimeoutException TimeoutId

newtype EventlogEvent

Wrapper for Eventlog events so they can be retrieved from the trace with selectTraceEventsDynamic.

Constructors

EventlogEvent String

newtype EventlogMarker

Wrapper for Eventlog markers so they can be retrieved from the trace with selectTraceEventsDynamic.

Constructors

EventlogMarker String

data ThreadId

type ThreadLabel = String

data Labelled a

Constructors

Labelled

Fields

l_labelled ∷ !a

l_label ∷ !(Maybe String)

data Trace a

Trace is a recursive data type, it is the trace of a IOSim computation. The trace will contain information about thread sheduling, blocking on TVars, and other internal state changes of IOSim. More importantly it also supports traces generated by the computation with say (which corresponds to using putStrLn in IO), traceEventM, or dynamically typed traces with traceM (which generalise the base library traceM)

See also: traceEvents, traceResult, selectTraceEvents, selectTraceEventsDynamic and printTraceEventsSay.

Constructors

Trace !Time !ThreadId !(Maybe ThreadLabel) !TraceEvent (Trace a)

TraceMainReturn !Time a ![Labelled ThreadId]

TraceMainException !Time SomeException ![Labelled ThreadId]

TraceDeadlock !Time ![Labelled ThreadId]

data TraceEvent

Constructors

EventSay String

EventLog Dynamic

EventThrow SomeException

EventThrowTo SomeException ThreadId

EventThrowToBlocked

EventThrowToWakeup

EventThrowToUnmasked (Labelled ThreadId)

EventThreadForked ThreadId

EventThreadFinished

EventThreadUnhandled SomeException

EventTxCommitted [Labelled TVarId] [TVarId]

EventTxAborted

EventTxBlocked [Labelled TVarId]

EventTxWakeup [Labelled TVarId]

EventTimerCreated TimeoutId TVarId Time

EventTimerUpdated TimeoutId Time

EventTimerCancelled TimeoutId

EventTimerExpired TimeoutId

liftST ∷ ST s a → IOSim s a

execReadTVar ∷ TVar s a → ST s a