Control.Monad.IOSim.Types

Documentation

newtype IOSim s a

Constructors

IOSim

Fields

unIOSim :: forall r. (a -> SimA s r) -> SimA s r

runIOSim :: IOSim s a -> SimA s a

traceM :: Typeable a => a -> IOSim s ()

traceSTM :: Typeable a => a -> STMSim s ()

liftST :: ST s a -> IOSim s a

data SimA s a where

Constructors

Return :: a -> SimA s a

Say :: String -> SimA s b -> SimA s b

Output :: Dynamic -> SimA s b -> SimA s b

LiftST :: ST s a -> (a -> SimA s b) -> SimA s b

GetMonoTime :: (Time -> SimA s b) -> SimA s b

GetWallTime :: (UTCTime -> SimA s b) -> SimA s b

SetWallTime :: UTCTime -> SimA s b -> SimA s b

UnshareClock :: SimA s b -> SimA s b

NewTimeout :: DiffTime -> (Timeout (IOSim s) -> SimA s b) -> SimA s b

UpdateTimeout :: Timeout (IOSim s) -> DiffTime -> SimA s b -> SimA s b

CancelTimeout :: Timeout (IOSim s) -> SimA s b -> SimA s b

Throw :: SomeException -> SimA s a

Catch :: Exception e => SimA s a -> (e -> SimA s a) -> (a -> SimA s b) -> SimA s b

Evaluate :: a -> (a -> SimA s b) -> SimA s b

Fork :: IOSim s () -> (ThreadId -> SimA s b) -> SimA s b

GetThreadId :: (ThreadId -> SimA s b) -> SimA s b

LabelThread :: ThreadId -> String -> SimA s b -> SimA s b

Atomically :: STM s a -> (a -> SimA s b) -> SimA s b

ThrowTo :: SomeException -> ThreadId -> SimA s a -> SimA s a

SetMaskState :: MaskingState -> IOSim s a -> (a -> SimA s b) -> SimA s b

GetMaskState :: (MaskingState -> SimA s b) -> SimA s b

YieldSim :: SimA s a -> SimA s a

ExploreRaces :: SimA s b -> SimA s b

Fix :: (x -> IOSim s x) -> (x -> SimA s r) -> SimA s r

type StepId = (ThreadId, Int)

type STMSim = STM

newtype STM s a

Constructors

STM

Fields

unSTM :: forall r. (a -> StmA s r) -> StmA s r

runSTM :: STM s a -> StmA s a

data StmA s a where

Constructors

ReturnStm :: a -> StmA s a

ThrowStm :: SomeException -> StmA s a

NewTVar :: Maybe String -> x -> (TVar s x -> StmA s b) -> StmA s b

LabelTVar :: String -> TVar s a -> StmA s b -> StmA s b

ReadTVar :: TVar s a -> (a -> StmA s b) -> StmA s b

WriteTVar :: TVar s a -> a -> StmA s b -> StmA s b

Retry :: StmA s b

OrElse :: StmA s a -> StmA s a -> (a -> StmA s b) -> StmA s b

SayStm :: String -> StmA s b -> StmA s b

OutputStm :: Dynamic -> StmA s b -> StmA s b

TraceTVar :: forall s a b. TVar s a -> (Maybe a -> a -> ST s TraceValue) -> StmA s b -> StmA s b

data StmTxResult s a

Constructors

StmTxCommitted
A committed transaction reports the vars that were written (in order of first write) so that the scheduler can unblock other threads that were blocked in STM transactions that read any of these vars.

It reports the vars that were read, so we can update vector clocks appropriately.

The third list of vars is ones that were created during this transaction. This is useful for an implementation of traceTVar.

It also includes the updated TVarId name supply.

Fields

a

[SomeTVar s]

written tvars

[SomeTVar s]

read tvars

[SomeTVar s]

created tvars

[Dynamic]

[String]

TVarId

StmTxBlocked [SomeTVar s]
A blocked transaction reports the vars that were read so that the scheduler can block the thread on those vars.

StmTxAborted [SomeTVar s] SomeException
An aborted transaction reports the vars that were read so that the vector clock can be updated.

data StmStack s b a where

Constructors

AtomicallyFrame :: StmStack s a a
Executing in the context of a top level atomically.

OrElseLeftFrame :: StmA s a -> (a -> StmA s b) -> Map TVarId (SomeTVar s) -> [SomeTVar s] -> [SomeTVar s] -> StmStack s b c -> StmStack s a c
Executing in the context of the left hand side of an orElse

OrElseRightFrame :: (a -> StmA s b) -> Map TVarId (SomeTVar s) -> [SomeTVar s] -> [SomeTVar s] -> StmStack s b c -> StmStack s a c
Executing in the context of the right hand side of an orElse

data family Timeout (m :: Type -> Type)

newtype TimeoutException

Constructors

TimeoutException TimeoutId

setCurrentTime :: UTCTime -> IOSim s ()

Set the current wall clock time for the thread's clock domain.

unshareClock :: IOSim s ()

Put the thread into a new wall clock domain, not shared with the parent thread. Changing the wall clock time in the new clock domain will not affect the other clock of other threads. All threads forked by this thread from this point onwards will share the new clock domain.

data ScheduleControl

Constructors

ControlDefault
default scheduling mode

ControlAwait [ScheduleMod]
if the current control is ControlAwait, the normal scheduling will proceed, until the thread found in the first ScheduleMod reaches the given step. At this point the thread is put to sleep, until after all the steps are followed.

ControlFollow [StepId] [ScheduleMod]
follow the steps then continue with schedule modifications. This control is set by followControl when controlTargets returns true.

data ScheduleMod

Constructors

ScheduleMod

Fields

scheduleModTarget :: StepId

Step at which the ScheduleMod is activated.

scheduleModControl :: ScheduleControl

ScheduleControl at the activation step. It is needed by extendScheduleControl when combining the discovered schedule with the initial one.

scheduleModInsertion :: [StepId]

Series of steps which are executed at the target step. This *includes* the target step, not necessarily as the last step.

data ExplorationOptions

Constructors

ExplorationOptions

Fields

explorationScheduleBound :: Int

explorationBranching :: Int

explorationStepTimelimit :: Maybe Int

explorationReplay :: Maybe ScheduleControl

type ExplorationSpec = ExplorationOptions -> ExplorationOptions

withScheduleBound :: Int -> ExplorationSpec

withBranching :: Int -> ExplorationSpec

withStepTimelimit :: Int -> ExplorationSpec

withReplay :: ScheduleControl -> ExplorationSpec

stdExplorationOptions :: ExplorationOptions

newtype EventlogEvent

Wrapper for Eventlog events so they can be retrieved from the trace with selectTraceEventsDynamic.

Constructors

EventlogEvent String

newtype EventlogMarker

Wrapper for Eventlog markers so they can be retrieved from the trace with selectTraceEventsDynamic.

Constructors

EventlogMarker String

data SimEventType

Constructors

EventSimStart ScheduleControl

EventSay String

EventLog Dynamic

EventMask MaskingState

EventThrow SomeException

EventThrowTo SomeException ThreadId

EventThrowToBlocked

EventThrowToWakeup

EventThrowToUnmasked (Labelled ThreadId)

EventThreadForked ThreadId

EventThreadFinished

EventThreadUnhandled SomeException

EventTxCommitted [Labelled TVarId] [Labelled TVarId] (Maybe Effect)

EventTxAborted (Maybe Effect)

EventTxBlocked [Labelled TVarId] (Maybe Effect)

EventTxWakeup [Labelled TVarId]

EventTimerCreated TimeoutId TVarId Time

EventTimerUpdated TimeoutId Time

EventTimerCancelled TimeoutId

EventTimerExpired TimeoutId

EventThreadSleep

EventThreadWake

EventDeschedule Deschedule

EventFollowControl ScheduleControl

EventAwaitControl StepId ScheduleControl

EventPerformAction StepId

EventReschedule ScheduleControl

EventUnblocked [ThreadId]

data SimEvent

Trace is a recursive data type, it is the trace of a IOSim computation. The trace will contain information about thread sheduling, blocking on TVars, and other internal state changes of IOSim. More importantly it also supports traces generated by the computation with say (which corresponds to using putStrLn in IO), traceEventM, or dynamically typed traces with traceM (which generalise the base library traceM)

It also contains information on races discovered.

See also: traceEvents, traceResult, selectTraceEvents, selectTraceEventsDynamic and printTraceEventsSay.

Constructors

SimEvent

Fields

seTime :: !Time

seThreadId :: !ThreadId

seThreadLabel :: !(Maybe ThreadLabel)

seType :: !SimEventType

SimPOREvent

Fields

seTime :: !Time

seThreadId :: !ThreadId

seStep :: !Int

seThreadLabel :: !(Maybe ThreadLabel)

seType :: !SimEventType

SimRacesFound [ScheduleControl]

data SimResult a

Constructors

MainReturn !Time a ![Labelled ThreadId]

MainException !Time SomeException ![Labelled ThreadId]

Deadlock !Time ![Labelled ThreadId]

Loop

type SimTrace a = Trace (SimResult a) SimEvent

data Trace a b where

A cons list with polymorphic nil, thus an octopus.

Trace Void a is an infinite stream

Trace () a is isomorphic to [a]

Usually used with a being a non empty sum type.

Bundled Patterns

pattern Trace :: Time -> ThreadId -> Maybe ThreadLabel -> SimEventType -> SimTrace a -> SimTrace a

Deprecated: Use SimTrace instead.

pattern SimTrace :: Time -> ThreadId -> Maybe ThreadLabel -> SimEventType -> SimTrace a -> SimTrace a

pattern SimPORTrace :: Time -> ThreadId -> Int -> Maybe ThreadLabel -> SimEventType -> SimTrace a -> SimTrace a

pattern TraceMainReturn :: Time -> a -> [Labelled ThreadId] -> SimTrace a

pattern TraceMainException :: Time -> SomeException -> [Labelled ThreadId] -> SimTrace a

pattern TraceDeadlock :: Time -> [Labelled ThreadId] -> SimTrace a

pattern TraceRacesFound :: [ScheduleControl] -> SimTrace a -> SimTrace a

pattern TraceLoop :: SimTrace a

ppTrace :: Show a => SimTrace a -> String

Pretty print simulation trace.

ppTrace_ :: SimTrace a -> String

Like ppTrace but does not show the result value.

ppSimEvent

Arguments

:: Int
width of the time

-> Int
width of thread id

-> Int
width of thread label

-> SimEvent

-> String

ppDebug :: SimTrace a -> x -> x

Trace each event using trace; this is useful when a trace ends with a pure error, e.g. an assertion.

type TraceEvent = SimEventType

Deprecated: Use SimEventType instead.

data Labelled a

Constructors

Labelled

Fields

l_labelled :: !a

l_label :: !(Maybe String)

data ThreadId

Constructors

RacyThreadId [Int]

ThreadId [Int]

childThreadId :: ThreadId -> Int -> ThreadId

setRacyThread :: ThreadId -> ThreadId

newtype TVarId

Constructors

TVarId Int

newtype TimeoutId

Constructors

TimeoutId Int

newtype ClockId

Constructors

ClockId [Int]

newtype VectorClock

Constructors

VectorClock

Fields

getVectorClock :: Map ThreadId Int

unTimeoutId :: TimeoutId -> Int

type ThreadLabel = String

type TVarLabel = String

data TVar s a

Constructors

TVar

Fields

tvarId :: !TVarId

The identifier of this var.

tvarLabel :: !(STRef s (Maybe TVarLabel))

Label.

tvarCurrent :: !(STRef s a)

The var's current value

tvarUndo :: !(STRef s [a])

A stack of undo values. This is only used while executing a transaction.

tvarBlocked :: !(STRef s ([ThreadId], Set ThreadId))

Thread Ids of threads blocked on a read of this var. It is represented in reverse order of thread wakeup, without duplicates.

To avoid duplicates efficiently, the operations rely on a copy of the thread Ids represented as a set.

tvarVClock :: !(STRef s VectorClock)

The vector clock of the current value.

tvarTrace :: !(STRef s (Maybe (Maybe a -> a -> ST s TraceValue)))

Callback to construct a trace which will be attached to the dynamic trace.

data SomeTVar s where

Constructors

SomeTVar :: !(TVar s a) -> SomeTVar s

data Deschedule

Constructors

Yield

Interruptable

Blocked

Terminated

Sleep

type SimM s = IOSim s

Deprecated: Use IOSim

type SimSTM = STM

Deprecated: Use STMSim