Control.Monad.IOSimPOR.Timeout

Documentation

data Timeout

An exception thrown to a thread by timeout to interrupt a timed-out computation.

timeout ∷ Int → IO a → IO (Maybe a)

unsafeTimeout ∷ Int → a → Maybe a

unsafeTimeout n a forces the evaluation of a, with a time limit of n microseconds.