Data.List.Trace

Documentation

data Trace a b

A cons list with polymorphic nil, thus an octopus.

Trace Void a is an infinite stream

Trace () a is isomorphic to [a]

Usually used with a being a non empty sum type.

Constructors

Cons b (Trace a b)

Nil a

ppTrace ∷ (a → String) → (b → String) → Trace a b → String

Pretty print an Trace.

toList ∷ Trace a b → [b]

fromList ∷ a → [b] → Trace a b

head ∷ Trace a b → b

tail ∷ Trace a b → Trace a b

filter ∷ (b → Bool) → Trace a b → Trace a b

length ∷ Trace a b → Int