Control.Monad.IOSim

Simulation monad

data IOSim s a

type STMSim = STM

IOSim's STM monad, as IOSim it is parametrised by s, e.g. STMSim s a is monadic expression of type a.

Run simulation

runSim :: forall a. (forall s. IOSim s a) -> Either Failure a

IOSim is a pure monad.

runSimOrThrow :: forall a. (forall s. IOSim s a) -> a

For quick experiments and tests it is often appropriate and convenient to simply throw failures as exceptions.

runSimStrictShutdown :: forall a. (forall s. IOSim s a) -> Either Failure a

Like runSim but fail when the main thread terminates if there are other threads still running or blocked. If one is trying to follow a strict thread clean-up policy then this helps testing for that.

data Failure

Simulation terminated a failure.

Constructors

FailureException SomeException
The main thread terminated with an exception.

FailureDeadlock ![Labelled IOSimThreadId]
The threads all deadlocked.

FailureSloppyShutdown [Labelled IOSimThreadId]
The main thread terminated normally but other threads were still alive, and strict shutdown checking was requested. See runSimStrictShutdown.

FailureEvaluation SomeException
An exception was thrown while evaluation the trace. This could be an internal assertion failure of `io-sim` or an unhandled exception in the simulation.

FailureInternal String
An internal failure of the simulator.

Please open an issue at https://github.com/input-output-hk/io-sim/issues.

runSimTrace :: forall a. (forall s. IOSim s a) -> SimTrace a

See runSimTraceST below.

runSimTraceST :: forall s a. IOSim s a -> ST s (SimTrace a)

The most general method of running IOSim is in the lazy ST monad. One can recover failures or the result from SimTrace with traceResult, or access SimEventTypes generated by the computation with traceEvents. A slightly more convenient way is exposed by runSimTrace.

QuickCheck Monadic combinators

monadicIOSim_ :: Testable a => (forall s. PropertyM (IOSim s) a) -> Property

Like https://hackage.haskell.org/package/QuickCheck-2.14.3/docs/Test-QuickCheck-Monadic.html#v:monadicST.

Note: it calls traceResult in non-strict mode, e.g. leaked threads do not cause failures.

Since: 1.4.1.0

monadicIOSim

Arguments

:: (Testable a, forall s. Monad (m s))

=> (SimTrace Property -> Property)
Allows to trace SimTrace in counterexamples. The simplest use case is to pass:

either (\e -> counterexample (show e) False) id . traceResult False

as monadicIOSim_ does.

-> (forall s a. m s a -> IOSim s a)
natural transformation from m to IOSim` s

-> (forall s. PropertyM (m s) a)

-> Property

A more general version of monadicIOSim_, which:

allows to run in monad stacks build on top of IOSim;

gives more control how to attach debugging information to failed tests.

Note, to use this combinator your monad needs to be defined as:

newtype M s a = M s { runM :: ReaderT State (IOSim s) a }

It's important that `M s` is a monad. For such a monad one you'll need provide a natural transformation: -- the state could also be created as an IOSim computation. nat :: forall s a. State -> M s a -> IOSim s a nat state m = runStateT (runM m) state

Since: 1.4.1.0

runIOSimGen :: (SimTrace a -> Property) -> (forall s. Gen (IOSim s a)) -> Gen Property

Like <runSTGen https://hackage.haskell.org/package/QuickCheck-2.14.3/docs/Test-QuickCheck-Monadic.html#v:runSTGen>.

Since: 1.4.1.0

Explore races using IOSimPOR

IOSimPOR is a different interpreter of IOSim which has the ability to discover race conditions and replay the simulation using a schedule which reverts them. For extended documentation how to use it see here.

IOSimPOR only discovers races between events which happen in the same time slot. In IOSim and IOSimPOR time only moves explicitly through timer events, e.g. things like threadDelay, registerDelay or the NonStandard API. The usual QuickCheck techniques can help explore different schedules of threads too.

exploreSimTrace

Arguments

:: forall a test. Testable test

=> (ExplorationOptions -> ExplorationOptions)
modify default exploration options

-> (forall s. IOSim s a)
a simulation to run

-> (Maybe (SimTrace a) -> SimTrace a -> test)
a callback which receives the previous trace (e.g. before reverting a race condition) and current trace

-> Property

Execute a simulation, discover & revert races. Note that this will execute the simulation multiple times with different schedules, and thus it's much more costly than a simple runSimTrace (also the simulation environments has much more state to track and hence it is slower).

On property failure it will show the failing schedule (ScheduleControl) which can be passed to controlSimTrace to reproduce the failure without discovering the schedule.

controlSimTrace

Arguments

:: forall a. Maybe Int
limit on the computation time allowed per scheduling step, for catching infinite loops etc.

-> ScheduleControl
a schedule to replay

note: must be either ControlDefault or ControlAwait.

-> (forall s. IOSim s a)
a simulation to run

-> SimTrace a

Run a simulation using a given schedule. This is useful to reproduce failing cases without exploring the races.

data ScheduleMod

A schedule modification inserted at given execution step.

Constructors

ScheduleMod

Fields

scheduleModTarget :: StepId

Step at which the ScheduleMod is activated.

scheduleModControl :: ScheduleControl

ScheduleControl at the activation step. It is needed by extendScheduleControl when combining the discovered schedule with the initial one.

scheduleModInsertion :: [StepId]

Series of steps which are executed at the target step. This *includes* the target step, not necessarily as the last step.

data ScheduleControl

Modified execution schedule.

Constructors

ControlDefault
default scheduling mode

ControlAwait [ScheduleMod]
if the current control is ControlAwait, the normal scheduling will proceed, until the thread found in the first ScheduleMod reaches the given step. At this point the thread is put to sleep, until after all the steps are followed.

ControlFollow [StepId] [ScheduleMod]
follow the steps then continue with schedule modifications. This control is set by followControl when controlTargets returns true.

Exploration options

type ExplorationSpec = ExplorationOptions -> ExplorationOptions

data ExplorationOptions

Race exploration options.

Constructors

ExplorationOptions

Fields

explorationScheduleBound :: Int

This is an upper bound on the number of schedules with race reversals that will be explored; a bound of zero means that the default schedule will be explored, but no others. Setting the bound to zero makes IOSimPOR behave rather like IOSim, in that only one schedule is explored, but (a) IOSimPOR is considerably slower, because it still collects information on potential races, and (b) the IOSimPOR schedule is different (based on priorities, in contrast to IOSim's round-robin), and plays better with shrinking.

The default value is `100`.

explorationBranching :: Int

The branching factor. This is the number of alternative schedules that IOSimPOR tries to run, per race reversal. With the default parameters, IOSimPOR will try to reverse the first 33 (100 div 3) races discovered using the default schedule, then (if 33 or more races are discovered), for each such reversed race, will run the reversal and try to reverse two more races in the resulting schedule. A high branching factor will explore more combinations of reversing fewer races, within the overall schedule bound. A branching factor of one will explore only schedules resulting from a single race reversal (unless there are fewer races available to be reversed than the schedule bound).

The default value is `3`.

explorationStepTimelimit :: Maybe Int

Limit on the computation time allowed per scheduling step, for catching infinite loops etc.

The default value is Nothing.

explorationReplay :: Maybe ScheduleControl

A schedule to replay.

The default value is Nothing.

explorationDebugLevel :: Int

Log detailed trace to stderr containing information on discovered races. The trace does not contain the result of the simulation, unless one will do that explicitly inside the simulation.

level 0: don't show any output, level 1: show simulation trace with discovered schedules level 2: show simulation trace with discovered schedules and races

NOTE: discovered schedules & races are not exposed to the user in the callback of exploreSimTrace or in the output of controlSimTrace.

stdExplorationOptions :: ExplorationOptions

withScheduleBound :: Int -> ExplorationSpec

withBranching :: Int -> ExplorationSpec

withStepTimelimit :: Int -> ExplorationSpec

withReplay :: ScheduleControl -> ExplorationSpec

Lift ST computations

liftST :: ST s a -> IOSim s a

Lift an ST computation to IOSim.

Note: you can use MonadST to lift ST computations, this is a more convenient function just for IOSim.

Simulation time

setCurrentTime :: UTCTime -> IOSim s ()

Set the current wall clock time for the thread's clock domain.

unshareClock :: IOSim s ()

Put the thread into a new wall clock domain, not shared with the parent thread. Changing the wall clock time in the new clock domain will not affect the other clock of other threads. All threads forked by this thread from this point onwards will share the new clock domain.

Simulation trace

type SimTrace a = Trace (SimResult a) SimEvent

A type alias for IOSim simulation trace. It comes with useful pattern synonyms.

data Trace a b

A cons list with polymorphic nil.

Trace Void a is an infinite stream

Trace () a is isomorphic to [a]

Usually used with a being a non empty sum type.

Constructors

Cons b (Trace a b)

Nil a

Bundled Patterns

pattern SimTrace :: Time -> IOSimThreadId -> Maybe ThreadLabel -> SimEventType -> SimTrace a -> SimTrace a

pattern SimPORTrace :: Time -> IOSimThreadId -> Int -> Maybe ThreadLabel -> SimEventType -> SimTrace a -> SimTrace a

pattern TraceDeadlock :: Time -> [Labelled IOSimThreadId] -> SimTrace a

pattern TraceLoop :: SimTrace a

pattern TraceMainReturn :: Time -> Labelled IOSimThreadId -> a -> [Labelled IOSimThreadId] -> SimTrace a

pattern TraceMainException :: Time -> Labelled IOSimThreadId -> SomeException -> [Labelled IOSimThreadId] -> SimTrace a

pattern TraceRacesFound :: [ScheduleControl] -> SimTrace a -> SimTrace a

pattern TraceInternalError :: String -> SimTrace a

data SimResult a

A result type of a simulation.

Constructors

MainReturn !Time !(Labelled IOSimThreadId) a ![Labelled IOSimThreadId]
Return value of the main thread.

MainException !Time !(Labelled IOSimThreadId) SomeException ![Labelled IOSimThreadId]
Exception thrown by the main thread.

Deadlock !Time ![Labelled IOSimThreadId]
Deadlock discovered in the simulation. Deadlocks are discovered if simply the simulation cannot do any progress in a given time slot and there's no event which would advance the time.

Loop
Only returned by IOSimPOR when a step execution took longer than explorationStepTimelimit was exceeded.

InternalError String
An IOSim bug, please report to https://github.com/input-output-hk/io-sim

data SimEvent

Trace is a recursive data type, it is the trace of a IOSim computation. The trace will contain information about thread scheduling, blocking on TVars, and other internal state changes of IOSim. More importantly it also supports traces generated by the computation with say (which corresponds to using putStrLn in IO), traceEventM, or dynamically typed traces with traceM (which generalise the base library traceM)

It also contains information on discovered races.

See also: traceEvents, traceResult, selectTraceEvents, selectTraceEventsDynamic and printTraceEventsSay.

Constructors

SimEvent
Used when using IOSim.

Fields

seTime :: !Time

seThreadId :: !IOSimThreadId

seThreadLabel :: !(Maybe ThreadLabel)

seType :: !SimEventType

SimPOREvent
Only used for IOSimPOR

Fields

seTime :: !Time

seThreadId :: !IOSimThreadId

seStep :: !Int

seThreadLabel :: !(Maybe ThreadLabel)

seType :: !SimEventType

SimRacesFound [ScheduleControl]
Only used for IOSimPOR

data SimEventType

Events recorded by the simulation.

Constructors

EventSay String
hold value of say

EventLog Dynamic
hold a dynamic value of traceM

EventMask MaskingState
masking state changed

EventThrow SomeException
throw exception

EventThrowTo SomeException IOSimThreadId
throw asynchronous exception (throwTo)

EventThrowToBlocked
the thread which executed throwTo is blocked

EventThrowToWakeup
the thread which executed throwTo is woken up

EventThrowToUnmasked (Labelled IOSimThreadId)
a target thread of throwTo unmasked its exceptions, this is paired with EventThrowToWakeup for threads which were blocked on throwTo

EventThreadForked IOSimThreadId
forked a thread

EventThreadFinished
thread terminated normally

EventThreadUnhandled SomeException
thread terminated by an unhandled exception

EventTxCommitted
committed STM transaction

Fields

[Labelled TVarId]

stm tx wrote to these

[Labelled TVarId]

and created these

(Maybe Effect)

effect performed (only for IOSimPOR) | aborted an STM transaction (by an exception)

For IOSimPOR it also holds performed effect.

EventTxAborted (Maybe Effect)

EventTxBlocked
STM transaction blocked (due to retry)

Fields

[Labelled TVarId]

(Maybe Effect)

effect performed (only for IOSimPOR)

EventTxWakeup [Labelled TVarId]
changed vars causing retry

EventUnblocked [IOSimThreadId]
unblocked threads by a committed STM transaction

EventThreadDelay TimeoutId Time
thread delayed

EventThreadDelayFired TimeoutId
thread woken up after a delay

EventTimeoutCreated TimeoutId IOSimThreadId Time
new timeout created (via timeout)

EventTimeoutFired TimeoutId
timeout fired

EventRegisterDelayCreated TimeoutId TVarId Time
registered delay (via registerDelay)

EventRegisterDelayFired TimeoutId
registered delay fired

EventTimerCreated TimeoutId TVarId Time
a new Timeout created (via newTimeout)

EventTimerCancelled TimeoutId
a Timeout was cancelled (via cancelTimeout)

EventTimerFired TimeoutId
a Timeout fired

EventThreadStatus
event traced when threadStatus is executed

Fields

IOSimThreadId

current thread

IOSimThreadId

queried thread

EventSimStart ScheduleControl
IOSimPOR event: new execution started exploring the given schedule.

EventThreadSleep
IOSimPOR event: the labelling thread was runnable, but its execution was delayed, until EventThreadWake.

Event inserted to mark a difference between a failed trace and a similar passing trace.

EventThreadWake
IOSimPOR event: marks when the thread was rescheduled by IOSimPOR

EventDeschedule Deschedule
IOSim and IOSimPOR event: a thread was descheduled

EventFollowControl ScheduleControl
IOSimPOR event: following given schedule

EventAwaitControl StepId ScheduleControl
IOSimPOR event: thread delayed to follow the given schedule

EventPerformAction StepId
IOSimPOR event: perform action of the given step

EventReschedule ScheduleControl
IOSimPOR event: reschedule a thread following the given ScheduleControl

EventEffect VectorClock Effect
IOSimPOR event: executed effect; Useful for debugging IOSimPOR or showing compact information about thread execution.

EventRaces Races
IOSimPOR event: races. Races are updated while we execute a simulation. Useful for debugging IOSimPOR.

type ThreadLabel = String

data IOSimThreadId

A thread id.

IOSimPOR: RacyThreadId indicates that this thread is taken into account when discovering races. A thread is marked as racy iff exploreRaces was executed in it or it's a thread forked by a racy thread.

Constructors

RacyThreadId [Int]
A racy thread (IOSimPOR only), shown in the trace with curly braces, e.g. `Thread {2,3}`.

ThreadId [Int]
A non racy thread. They have higher priority than racy threads in IOSimPOR scheduler.

data Labelled a

A labelled value.

For example labelThread or labelTVar will insert a label to IOSimThreadId (or TVarId).

Constructors

Labelled

Fields

l_labelled :: !a

l_label :: !(Maybe String)

Dynamic Tracing

traceM :: Typeable a => a -> IOSim s ()

IOSim has the ability to store any Typeable value in its trace which can then be recovered with selectTraceEventsDynamic or selectTraceEventsDynamic'.

traceSTM :: Typeable a => a -> STMSim s ()

Trace a value, in the same was as traceM does, but from the STM monad. This is primarily useful for debugging.

Pretty printers

ppTrace :: Show a => SimTrace a -> String

Pretty print simulation trace.

Note: this is not a streaming function, it will evaluate the whole trace before printing it. If you need to print a very large trace, you might want to use

ppTrace show (ppSimEvent 0 0 0)

ppTrace_ :: SimTrace a -> String

Like ppTrace but does not show the result value.

ppEvents :: [(Time, IOSimThreadId, Maybe ThreadLabel, SimEventType)] -> String

Pretty print a timestamped event.

ppSimEvent

Arguments

:: Int
width of the time

-> Int
width of thread id

-> Int
width of thread label

-> SimEvent

-> String

Pretty print a SimEvent.

ppDebug :: SimTrace a -> x -> x

Trace each event using trace; this is useful when a trace ends with a pure error, e.g. an assertion.

Selectors

traceEvents :: SimTrace a -> [(Time, IOSimThreadId, Maybe ThreadLabel, SimEventType)]

Turn SimTrace into a list of timestamped events.

traceResult

Arguments

:: Bool
if True the simulation will fail if there are any threads which didn't terminated when the main thread terminated.

-> SimTrace a
simulation trace

-> Either Failure a

Fold through the trace and return either Failure or a simulation result, i.e. the return value of the main thread.

list selectors

selectTraceEvents :: (Time -> SimEventType -> Maybe b) -> SimTrace a -> [b]

Select events according to the predicate function. It throws an error if the simulation ends with Failure.

selectTraceEvents' :: (Time -> SimEventType -> Maybe b) -> SimTrace a -> [b]

Like selectTraceEvents, but it returns even if the simulation trace ends with Failure.

selectTraceEventsDynamic :: forall a b. Typeable b => SimTrace a -> [b]

Select all the traced values matching the expected type. It relies on the sim's dynamic trace facility.

For convenience, it throws exceptions for abnormal sim termination.

selectTraceEventsDynamicWithTime :: forall a b. Typeable b => SimTrace a -> [(Time, b)]

Like selectTraceEventsDynamic but it also captures time of the trace event.

selectTraceEventsDynamic' :: forall a b. Typeable b => SimTrace a -> [b]

Like selectTraceEventsDynamic but it returns even if the simulation trace ends with Failure.

selectTraceEventsDynamicWithTime' :: forall a b. Typeable b => SimTrace a -> [(Time, b)]

Like selectTraceEventsDynamic' but it also captures time of the trace event.

selectTraceEventsSay :: SimTrace a -> [String]

Get a trace of EventSay.

For convenience, it throws exceptions for abnormal sim termination.

selectTraceEventsSayWithTime :: SimTrace a -> [(Time, String)]

Like selectTraceEventsSay but it also captures time of the trace event.

selectTraceEventsSay' :: SimTrace a -> [String]

Like selectTraceEventsSay but it returns even if the simulation trace ends with Failure.

selectTraceEventsSayWithTime' :: SimTrace a -> [(Time, String)]

Like selectTraceEventsSay' but it also captures time of the trace event.

selectTraceRaces :: SimTrace a -> [ScheduleControl]

trace selectors

traceSelectTraceEvents :: (Time -> SimEventType -> Maybe b) -> Trace a SimEvent -> Trace a b

The most general select function. It is a total function.

traceSelectTraceEventsDynamic :: forall a b. Typeable b => Trace a SimEvent -> Trace a b

Select dynamic events. It is a total function.

traceSelectTraceEventsSay :: forall a. Trace a SimEvent -> Trace a String

Select say events. It is a total function.

IO printer

printTraceEventsSay :: SimTrace a -> IO ()

Print all EventSay to the console.

For convenience, it throws exceptions for abnormal sim termination.

Eventlog

newtype EventlogEvent

Wrapper for Eventlog events so they can be retrieved from the trace with selectTraceEventsDynamic.

Constructors

EventlogEvent String

newtype EventlogMarker

Wrapper for Eventlog markers so they can be retrieved from the trace with selectTraceEventsDynamic.

Constructors

EventlogMarker String

Low-level API

data Timeout s

newTimeout :: DiffTime -> IOSim s (Timeout s)

readTimeout :: Timeout s -> STM s TimeoutState

cancelTimeout :: Timeout s -> IOSim s ()

awaitTimeout :: Timeout s -> STM s Bool