Control.Monad.IOSim

Simulation monad

data SimM s a

type SimSTM = STM

Run simulation

runSim ∷ ∀ a. (∀ s. SimM s a) → Either Failure a

runSimOrThrow ∷ ∀ a. (∀ s. SimM s a) → a

For quick experiments and tests it is often appropriate and convenient to simply throw failures as exceptions.

runSimStrictShutdown ∷ ∀ a. (∀ s. SimM s a) → Either Failure a

Like runSim but also fail if when the main thread terminates, there are other threads still running or blocked. If one is trying to follow a strict thread cleanup policy then this helps testing for that.

data Failure

Simulation termination with failure

Constructors

FailureException SomeException
The main thread terminated with an exception

FailureDeadlock
The threads all deadlocked

FailureSloppyShutdown [ThreadId]
The main thread terminated normally but other threads were still alive, and strict shutdown checking was requested. See runSimStrictShutdown

runSimTrace ∷ ∀ a. (∀ s. SimM s a) → Trace a

runSimTraceST ∷ ∀ s a. SimM s a → ST s (Trace a)

liftST ∷ ST s a → SimM s a

traceM ∷ Typeable a ⇒ a → SimM s ()

Simulation time

setCurrentTime ∷ UTCTime → SimM s ()

Set the current wall clock time for the thread's clock domain.

unshareClock ∷ SimM s ()

Put the thread into a new wall clock domain, not shared with the parent thread. Changing the wall clock time in the new clock domain will not affect the other clock of other threads. All threads forked by this thread from this point onwards will share the new clock domain.

Simulation trace

data Trace a

Constructors

Trace !Time !ThreadId !(Maybe ThreadLabel) !TraceEvent (Trace a)

TraceMainReturn !Time a ![ThreadId]

TraceMainException !Time SomeException ![ThreadId]

TraceDeadlock !Time ![ThreadId]

data TraceEvent

Constructors

EventSay String

EventLog Dynamic

EventThrow SomeException

EventThrowTo SomeException ThreadId

EventThrowToBlocked

EventThrowToWakeup

EventThrowToUnmasked ThreadId

EventThreadForked ThreadId

EventThreadFinished

EventThreadUnhandled SomeException

EventTxCommitted [TVarId] [TVarId]

EventTxAborted

EventTxBlocked [TVarId]

EventTxWakeup [TVarId]

EventTimerCreated TimeoutId TVarId Time

EventTimerUpdated TimeoutId Time

EventTimerCancelled TimeoutId

EventTimerExpired TimeoutId

traceEvents ∷ Trace a → [(Time, ThreadId, Maybe ThreadLabel, TraceEvent)]

traceResult ∷ Bool → Trace a → Either Failure a

selectTraceEvents ∷ (TraceEvent → Maybe b) → Trace a → [b]

selectTraceEventsDynamic ∷ ∀ a b. Typeable b ⇒ Trace a → [b]

Select all the traced values matching the expected type. This relies on the sim's dynamic trace facility.

For convenience, this throws exceptions for abnormal sim termination.

selectTraceEventsSay ∷ Trace a → [String]

Get a trace of EventSay.

For convenience, this throws exceptions for abnormal sim termination.

printTraceEventsSay ∷ Trace a → IO ()

Print all EventSay to the console.

For convenience, this throws exceptions for abnormal sim termination.

Eventlog

newtype EventlogEvent

Wrapper for Eventlog events so they can be retrieved from the trace with selectTraceEventsDynamic.

Constructors

EventlogEvent String

newtype EventlogMarker

Wrapper for Eventlog markers so they can be retrieved from the trace with selectTraceEventsDynamic.

Constructors

EventlogMarker String

Low-level API

execReadTVar ∷ TVar s a → ST s a