Data.List.Trace

Documentation

data Trace a b

A cons list with polymorphic nil.

Trace Void a is an infinite stream

Trace () a is isomorphic to [a]

Usually used with a being a non empty sum type.

Constructors

Cons b (Trace a b)

Nil a

ppTrace :: (a -> String) -> (b -> String) -> Trace a b -> String

Pretty print a Trace.

toList :: Trace a b -> [b]

fromList :: a -> [b] -> Trace a b

head :: Trace a b -> b

tail :: Trace a b -> Trace a b

filter :: (b -> Bool) -> Trace a b -> Trace a b

length :: Trace a b -> Int

take :: Int -> Trace a b -> Trace (Maybe a) b

Take the first n elements of a Trace, converting each to ().

takeWhile :: (b -> Bool) -> Trace a b -> Trace (Maybe a) b

Take elements from the Trace while the predicate holds, converting each to ().

drop :: Int -> Trace a b -> Trace a b

Drop the first n elements of a Trace.

dropWhile :: (b -> Bool) -> Trace a b -> Trace a b

Drop elements from the Trace while the predicate holds.